+++ packages/core/src/agent/canisterStatus/index.ts
@@ -127,7 +127,6 @@ export const request = async (options: CanisterStatusOptions): Promise<StatusMap
         const { path, data } = lookup(certificate, uniquePaths[index]);
         if (!data) {
           // Typically, the cert lookup will throw
-          console.warn(`Expected to find result for path ${path}, but instead found nothing.`);
           if (typeof path === 'string') {
             status.set(path, null);
           } else {
@@ -177,10 +176,6 @@ export const request = async (options: CanisterStatusOptions): Promise<StatusMap
         } else {
           status.set(path, null);
         }
-        console.group();
-        console.warn(`Expected to find result for path ${path}, but instead found nothing.`);
-        console.warn(error);
-        console.groupEnd();
       }
     })();
   });