+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - feat(agent): introduce the `SubnetStatus` utility namespace to request subnet information directly from the IC public API
 - feat(agent): export `IC_STATE_ROOT_DOMAIN_SEPARATOR` constant
 - fix(agent): check if canister is in ranges for certificates without delegation
+- fix(agent): verify all query signatures instead of only the first one
 - refactor(agent): only declare IC URLs once in the `HttpAgent` class
 - refactor(agent): split inner logic of `check_canister_ranges` into functions
 - test(principal): remove unneeded dependency